create better file structure, organise code
