+++ src/components/Game.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-  import { PLAYERS } from "./constants";
+  import { PLAYERS } from "../logic/constants";
   import Grid from "./Grid.svelte";
   import Scoreboard from "./Scoreboard.svelte";
-  import type { PlayerId, WinEvent } from "./types";
+  import type { PlayerId, WinEvent } from "../logic/types";
 
   let starter: PlayerId = Math.floor(Math.random() * PLAYERS);
   let turn: PlayerId = starter;
