+++ src/components/ModeSwitcher.svelte
