+++ src/logic/winning.ts
